+++ src/test_data_tools/test_models.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
-__all__ = ['DoseResponseTestCase', 'LassoTestCase']
+__all__ = ['DoseResponseTestCase', 'LassoTestCase', 'LinearTestCase',
+           'PowerLawTestCase']
 
 import unittest
 
@@ -33,8 +34,39 @@ class DoseResponseTestCase(unittest.TestCase):
     def test_ec(self):
         self.assertEqual(round(1 - self.model.ec() / self.params[0], 0), 0)
 
+
 class LassoTestCase(unittest.TestCase):
     @unittest.skip('** NOTE **: data_tools.models.Lasso test unit is not yet'
                    + ' implemented.')
     def test_null(self):
         pass
+
+
+class LinearTestCase(unittest.TestCase):
+    def setUp(self):
+        # Fit y = x
+        self.x = [0, 1, 2, 3, 4, 5]
+        self.y = [0, 1, 2, 3, 4, 5]
+
+        self.model = models.Linear(self.x, self.y)
+
+    def test_intercept(self):
+        self.assertEqual(self.model.intercept, 0.0)
+
+    def test_slope(self):
+        self.assertEqual(self.model.slope, 1.0)
+
+
+class PowerLawTestCase(unittest.TestCase):
+    def setUp(self):
+        # Fit y = x^-2
+        self.x = [1, 2, 4]
+        self.y = [1, 0.25, 0.0625]
+
+        self.model = models.PowerLaw(self.x, self.y)
+
+    def test_exponent(self):
+        self.assertEqual(self.model.k, -2.0)
+
+    def test_const(self):
+        self.assertEqual(self.model.a, 1.0)